Device snapshot processed: pedometer, rolls and battery charge. Alert on lost link added.

+++ src/main/java/ru/mecotrade/kidtracker/controller/model/Position.java
@@ -21,7 +21,7 @@ public class Position {
 
     private double accuracy;
 
-    private double battery;
+    private int battery;
 
     private int pedometer;
 
@@ -30,5 +30,4 @@ public class Position {
     private boolean lowBattery;
 
     private boolean sos;
-
 }
